Client: Inline the command definitions
