--- client/cli/src/cli-device-scan.ts
@@ -1,13 +1,0 @@
-import { Command } from "commander";
-
-
-const program = new Command();
-
-program
-    .action(async () => {
-        console.log("I'm scanning");
-    });
-
-program
-    .showHelpAfterError()
-    .parse();
