+++ client/cli/src/cli-device.ts
@@ -4,6 +4,12 @@ import { Command } from "commander";
 const program = new Command();
 
 program
-    .command("scan", "Scan for device advertisements")
+    .command("scan")
+    .description("Scan for device advertisements")
+    .action(async () => {
+        console.log("I'm scanning");
+    });
+
+program
     .showHelpAfterError()
     .parse();
