+++ client/cli/src/cli-module.ts
@@ -8,8 +8,32 @@ import { DefaultService as Api } from "../generatedApiClient";
 const program = new Command();
 
 program
-    .command("create", "Create a new module")
-    .command("desc", "Describe an existing module");
+    .command("create")
+    .description("Create a new module")
+    .argument("<module-name-string>", "Name to give to module")
+    .argument("<input-file>", "Path to module's .wasm file")
+    .action(async (name, wasmPath) => {
+        const wasm = await readFile(wasmPath);
+        const wasmBlob = new Blob([wasm], { type: "application/wasm" });
+        const result = await Api.postFileModule({
+            name, wasm: wasmBlob
+        });
+        console.log(JSON.stringify(result, null, 4));
+    });
+
+program
+    .command("desc")
+    .description("Describe an existing module")
+    .argument("<module-id-string>", "ID of the module")
+    .argument("<description-file>", "Path to JSON file describing functions of the module")
+    .action(async (id, descPath) => {
+        const descObj = JSON.parse(
+            await readFile(descPath, "utf8")
+        );
+        const result = await Api.postFileModuleUpload(id, descObj);
+        console.log(JSON.stringify(result, null, 4));
+    });
+
 
 program
     .command("show")
